refactor(global): clean code
