+++ package.json
@@ -42,8 +42,8 @@
     "@apollo/client": "^3.1.1",
     "@babel/core": "~7.11.6",
     "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/preset-env": "^7.11.5",
     "@babel/plugin-transform-runtime": "^7.12.0",
+    "@babel/preset-env": "^7.11.5",
     "@babel/preset-typescript": "^7.10.4",
     "@bem-react/classname": "^1.5.8",
     "@bem-react/classnames": "^1.3.9",
