+++ src/components/ExcelTable/Cells/index.ts
@@ -1,0 +1,3 @@
+import CellWithError from './CellWithError';
+
+export { CellWithError };
