+++ src/pages/Scheme/components/Table/Table.tsx
@@ -6,18 +6,11 @@ import { SelectedCell, TableEntities } from 'components/ExcelTable/types';
 import { isEmpty } from 'fp-ts/Array';
 import tableDuck from 'store/tableDuck';
 import { RootState } from 'store/types';
-import { Nullable, ProjectStructure } from 'types';
+import { Nullable } from 'types';
 import { mockTableRows, unpackData } from 'utils';
 
 import { GET_TABLE_TEMPLATE } from './queries';
-
-export interface TemplateProjectData {
-  project: {
-    template: {
-      structure: ProjectStructure;
-    };
-  };
-}
+import { TemplateProjectData } from './types';
 
 interface IProps {
   onSelect?: (data: Nullable<SelectedCell>) => void;
