+++ src/pages/Scheme/components/Table/types.ts
@@ -1,0 +1,9 @@
+import { ProjectStructure } from 'types';
+
+export interface TemplateProjectData {
+  project: {
+    template: {
+      structure: ProjectStructure;
+    };
+  };
+}
